merge upstream

+++ classes/journal/SectionDAO.inc.php
@@ -15,10 +15,16 @@
  * @brief Operations for retrieving and modifying Section objects.
  */
 
-use PKP\submission\PKPSubmission;
-use PKP\context\PKPSectionDAO;
+namespace APP\journal;
 
-import('classes.journal.Section');
+use APP\core\Services;
+use APP\facades\Repo;
+use PKP\cache\CacheManager;
+use PKP\context\PKPSectionDAO;
+use PKP\db\DAORegistry;
+use PKP\db\DAOResultFactory;
+use PKP\plugins\HookRegistry;
+use PKP\submission\PKPSubmission;
 
 class SectionDAO extends PKPSectionDAO
 {
@@ -320,12 +326,15 @@ class SectionDAO extends PKPSectionDAO
      */
     public function deleteById($sectionId, $contextId = null)
     {
+        // No articles should exist in this section
+        $collector = Repo::submission()->getCollector()->filterBySectionIds([(int) $sectionId]);
+        $count = Repo::submission()->getCount($collector);
+        if ($count) {
+            throw new Exception('Tried to delete a section that has one or more submissions assigned to it.');
+        }
+
         $subEditorsDao = DAORegistry::getDAO('SubEditorsDAO'); /* @var $subEditorsDao SubEditorsDAO */
         $subEditorsDao->deleteBySubmissionGroupId($sectionId, ASSOC_TYPE_SECTION, $contextId);
-
-        // Remove articles from this section
-        $submissionDao = DAORegistry::getDAO('SubmissionDAO'); /* @var $submissionDao SubmissionDAO */
-        $submissionDao->removeSubmissionsFromSection($sectionId);
 
         if (isset($contextId) && !$this->sectionExists($sectionId, $contextId)) {
             return false;
@@ -388,13 +397,15 @@ class SectionDAO extends PKPSectionDAO
         if (!$issue->getPublished()) {
             $allowedStatuses[] = PKPSubmission::STATUS_SCHEDULED;
         }
-        $submissionsIterator = Services::get('submission')->getMany([
-            'contextId' => $issue->getJournalId(),
-            'issueIds' => $issueId,
-            'status' => $allowedStatuses,
-        ]);
+        $collector = Repo::submission()->getCollector();
+        $collector
+            ->filterByContextIds([$issue->getJournalId()])
+            ->filterByIssueIds([$issueId])
+            ->filterByStatus($allowedStatuses)
+            ->orderBy($collector::ORDERBY_SEQUENCE, $collector::ORDER_DIR_ASC);
+        $submissions = Repo::submission()->getMany($collector);
         $sectionIds = [];
-        foreach ($submissionsIterator as $submission) {
+        foreach ($submissions as $submission) {
             $sectionIds[] = $submission->getCurrentPublication()->getData('sectionId');
         }
         if (empty($sectionIds)) {
@@ -658,4 +669,8 @@ class SectionDAO extends PKPSectionDAO
             [(float) $seq, (int) $issueId, (int) $sectionId]
         );
     }
+}
+
+if (!PKP_STRICT_MODE) {
+    class_alias('\APP\journal\SectionDAO', '\SectionDAO');
 }
